Fix changes missed in db folder move
